+++ lib/php/autoload.php
@@ -4,4 +4,4 @@ require_once(__DIR__ . '/src/RulesetInterface.php');
 require_once(__DIR__ . '/src/ClientInterface.php');
 require_once(__DIR__ . '/src/Client.php');
 require_once(__DIR__ . '/src/Ruleset.php');
-require_once(__DIR__ . '/src/Emojione.php');
+require_once(__DIR__ . '/src/JoyPixels.php');