+++ lib/php/test/ConversionTest.php
@@ -4,9 +4,9 @@
 Tests based on lib/tests.md
 */
 
-namespace Emojione\Test;
+namespace JoyPixels\Test;
 
-use Emojione\Emojione;
+use JoyPixels\JoyPixels;
 
 class ConversionTest extends \PHPUnit_Framework_TestCase
 {
@@ -22,15 +22,15 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
     {
         $unicode   = '🐌';
         $shortname = ':snail:';
-        $image     = '<img class="emojione" alt="🐌" title=":snail:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f40c.png"/>';
-        $image_fix = '<img class="emojione" alt="&#x1f40c;" title=":snail:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f40c.png"/>';
+        $image     = '<img class="joypixels" alt="🐌" title=":snail:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f40c.png"/>';
+        $image_fix = '<img class="joypixels" alt="&#x1f40c;" title=":snail:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f40c.png"/>';
 
-        $this->assertEquals(Emojione::toShort($unicode), $shortname);
-        $this->assertEquals(Emojione::shortnameToImage($shortname), $image_fix);
-        $this->assertEquals(Emojione::shortnameToUnicode($shortname), $unicode);
-        $this->assertEquals(Emojione::unicodeToImage($unicode), $image);
-        $this->assertEquals(Emojione::toImage($unicode), $image);
-        $this->assertEquals(Emojione::toImage($shortname), $image_fix);
+        $this->assertEquals(JoyPixels::toShort($unicode), $shortname);
+        $this->assertEquals(JoyPixels::shortnameToImage($shortname), $image_fix);
+        $this->assertEquals(JoyPixels::shortnameToUnicode($shortname), $unicode);
+        $this->assertEquals(JoyPixels::unicodeToImage($unicode), $image);
+        $this->assertEquals(JoyPixels::toImage($unicode), $image);
+        $this->assertEquals(JoyPixels::toImage($shortname), $image_fix);
     }
 
     /**
@@ -40,17 +40,17 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
      */
     public function testShortnameInsideSentence()
     {
-        $unicode   = 'The 🦄 is EmojiOne\'s official mascot.';
-        $shortname = 'The :unicorn: is EmojiOne\'s official mascot.';
-        $image     = 'The <img class="emojione" alt="🦄" title=":unicorn:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f984.png"/> is EmojiOne\'s official mascot.';
-        $image_fix = 'The <img class="emojione" alt="&#x1f984;" title=":unicorn:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f984.png"/> is EmojiOne\'s official mascot.';
+        $unicode   = 'The 🦄 was EmojiOne\'s official mascot.';
+        $shortname = 'The :unicorn: was EmojiOne\'s official mascot.';
+        $image     = 'The <img class="joypixels" alt="🦄" title=":unicorn:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f984.png"/> was EmojiOne\'s official mascot.';
+        $image_fix = 'The <img class="joypixels" alt="&#x1f984;" title=":unicorn:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f984.png"/> was EmojiOne\'s official mascot.';
 
-        $this->assertEquals(Emojione::toShort($unicode), $shortname);
-        $this->assertEquals(Emojione::shortnameToImage($shortname), $image_fix);
-        $this->assertEquals(Emojione::shortnameToUnicode($shortname), $unicode);
-        $this->assertEquals(Emojione::unicodeToImage($unicode), $image);
-        $this->assertEquals(Emojione::toImage($unicode), $image);
-        $this->assertEquals(Emojione::toImage($shortname), $image_fix);
+        $this->assertEquals(JoyPixels::toShort($unicode), $shortname);
+        $this->assertEquals(JoyPixels::shortnameToImage($shortname), $image_fix);
+        $this->assertEquals(JoyPixels::shortnameToUnicode($shortname), $unicode);
+        $this->assertEquals(JoyPixels::unicodeToImage($unicode), $image);
+        $this->assertEquals(JoyPixels::toImage($unicode), $image);
+        $this->assertEquals(JoyPixels::toImage($shortname), $image_fix);
     }
 
     /**
@@ -60,17 +60,17 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
      */
     public function testShortnameInsideSentenceWithComma()
     {
-        $unicode   = 'The 🦄, is EmojiOne\'s official mascot.';
-        $shortname = 'The :unicorn:, is EmojiOne\'s official mascot.';
-        $image     = 'The <img class="emojione" alt="🦄" title=":unicorn:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f984.png"/>, is EmojiOne\'s official mascot.';
-        $image_fix = 'The <img class="emojione" alt="&#x1f984;" title=":unicorn:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f984.png"/>, is EmojiOne\'s official mascot.';
+        $unicode   = 'The 🦄, was EmojiOne\'s official mascot.';
+        $shortname = 'The :unicorn:, was EmojiOne\'s official mascot.';
+        $image     = 'The <img class="joypixels" alt="🦄" title=":unicorn:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f984.png"/>, was EmojiOne\'s official mascot.';
+        $image_fix = 'The <img class="joypixels" alt="&#x1f984;" title=":unicorn:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f984.png"/>, was EmojiOne\'s official mascot.';
 
-        $this->assertEquals(Emojione::toShort($unicode), $shortname);
-        $this->assertEquals(Emojione::shortnameToImage($shortname), $image_fix);
-        $this->assertEquals(Emojione::shortnameToUnicode($shortname), $unicode);
-        $this->assertEquals(Emojione::unicodeToImage($unicode), $image);
-        $this->assertEquals(Emojione::toImage($unicode), $image);
-        $this->assertEquals(Emojione::toImage($shortname), $image_fix);
+        $this->assertEquals(JoyPixels::toShort($unicode), $shortname);
+        $this->assertEquals(JoyPixels::shortnameToImage($shortname), $image_fix);
+        $this->assertEquals(JoyPixels::shortnameToUnicode($shortname), $unicode);
+        $this->assertEquals(JoyPixels::unicodeToImage($unicode), $image);
+        $this->assertEquals(JoyPixels::toImage($unicode), $image);
+        $this->assertEquals(JoyPixels::toImage($shortname), $image_fix);
     }
 
     /**
@@ -82,15 +82,15 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
     {
         $unicode   = '🐌 mail.';
         $shortname = ':snail: mail.';
-        $image     = '<img class="emojione" alt="🐌" title=":snail:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f40c.png"/> mail.';
-        $image_fix = '<img class="emojione" alt="&#x1f40c;" title=":snail:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f40c.png"/> mail.';
+        $image     = '<img class="joypixels" alt="🐌" title=":snail:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f40c.png"/> mail.';
+        $image_fix = '<img class="joypixels" alt="&#x1f40c;" title=":snail:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f40c.png"/> mail.';
 
-        $this->assertEquals(Emojione::toShort($unicode), $shortname);
-        $this->assertEquals(Emojione::shortnameToImage($shortname), $image_fix);
-        $this->assertEquals(Emojione::shortnameToUnicode($shortname), $unicode);
-        $this->assertEquals(Emojione::unicodeToImage($unicode), $image);
-        $this->assertEquals(Emojione::toImage($unicode), $image);
-        $this->assertEquals(Emojione::toImage($shortname), $image_fix);
+        $this->assertEquals(JoyPixels::toShort($unicode), $shortname);
+        $this->assertEquals(JoyPixels::shortnameToImage($shortname), $image_fix);
+        $this->assertEquals(JoyPixels::shortnameToUnicode($shortname), $unicode);
+        $this->assertEquals(JoyPixels::unicodeToImage($unicode), $image);
+        $this->assertEquals(JoyPixels::toImage($unicode), $image);
+        $this->assertEquals(JoyPixels::toImage($shortname), $image_fix);
     }
 
     /**
@@ -102,15 +102,15 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
     {
         $unicode   = '🐌\'s are cool!';
         $shortname = ':snail:\'s are cool!';
-        $image     = '<img class="emojione" alt="🐌" title=":snail:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f40c.png"/>\'s are cool!';
-        $image_fix = '<img class="emojione" alt="&#x1f40c;" title=":snail:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f40c.png"/>\'s are cool!';
+        $image     = '<img class="joypixels" alt="🐌" title=":snail:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f40c.png"/>\'s are cool!';
+        $image_fix = '<img class="joypixels" alt="&#x1f40c;" title=":snail:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f40c.png"/>\'s are cool!';
 
-        $this->assertEquals(Emojione::toShort($unicode), $shortname);
-        $this->assertEquals(Emojione::shortnameToImage($shortname), $image_fix);
-        $this->assertEquals(Emojione::shortnameToUnicode($shortname), $unicode);
-        $this->assertEquals(Emojione::unicodeToImage($unicode), $image);
-        $this->assertEquals(Emojione::toImage($unicode), $image);
-        $this->assertEquals(Emojione::toImage($shortname), $image_fix);
+        $this->assertEquals(JoyPixels::toShort($unicode), $shortname);
+        $this->assertEquals(JoyPixels::shortnameToImage($shortname), $image_fix);
+        $this->assertEquals(JoyPixels::shortnameToUnicode($shortname), $unicode);
+        $this->assertEquals(JoyPixels::unicodeToImage($unicode), $image);
+        $this->assertEquals(JoyPixels::toImage($unicode), $image);
+        $this->assertEquals(JoyPixels::toImage($shortname), $image_fix);
     }
 
     /**
@@ -120,17 +120,17 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
      */
     public function testShortnameAtEndOfSentence()
     {
-        $unicode   = 'EmojiOne\'s official mascot is 🦄.';
-        $shortname = 'EmojiOne\'s official mascot is :unicorn:.';
-        $image     = 'EmojiOne\'s official mascot is <img class="emojione" alt="🦄" title=":unicorn:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f984.png"/>.';
-        $image_fix = 'EmojiOne\'s official mascot is <img class="emojione" alt="&#x1f984;" title=":unicorn:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f984.png"/>.';
+        $unicode   = 'EmojiOne\'s official mascot was 🦄.';
+        $shortname = 'EmojiOne\'s official mascot was :unicorn:.';
+        $image     = 'EmojiOne\'s official mascot was <img class="joypixels" alt="🦄" title=":unicorn:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f984.png"/>.';
+        $image_fix = 'EmojiOne\'s official mascot was <img class="joypixels" alt="&#x1f984;" title=":unicorn:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f984.png"/>.';
 
-        $this->assertEquals(Emojione::toShort($unicode), $shortname);
-        $this->assertEquals(Emojione::shortnameToImage($shortname), $image_fix);
-        $this->assertEquals(Emojione::shortnameToUnicode($shortname), $unicode);
-        $this->assertEquals(Emojione::unicodeToImage($unicode), $image);
-        $this->assertEquals(Emojione::toImage($unicode), $image);
-        $this->assertEquals(Emojione::toImage($shortname), $image_fix);
+        $this->assertEquals(JoyPixels::toShort($unicode), $shortname);
+        $this->assertEquals(JoyPixels::shortnameToImage($shortname), $image_fix);
+        $this->assertEquals(JoyPixels::shortnameToUnicode($shortname), $unicode);
+        $this->assertEquals(JoyPixels::unicodeToImage($unicode), $image);
+        $this->assertEquals(JoyPixels::toImage($unicode), $image);
+        $this->assertEquals(JoyPixels::toImage($shortname), $image_fix);
     }
 
     /**
@@ -140,17 +140,17 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
      */
     public function testShortnameAtEndOfSentenceWithAlternatePunctuation()
     {
-        $unicode   = 'EmojiOne\'s official mascot is 🦄!';
-        $shortname = 'EmojiOne\'s official mascot is :unicorn:!';
-        $image     = 'EmojiOne\'s official mascot is <img class="emojione" alt="🦄" title=":unicorn:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f984.png"/>!';
-        $image_fix = 'EmojiOne\'s official mascot is <img class="emojione" alt="&#x1f984;" title=":unicorn:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f984.png"/>!';
+        $unicode   = 'EmojiOne\'s official mascot was 🦄!';
+        $shortname = 'EmojiOne\'s official mascot was :unicorn:!';
+        $image     = 'EmojiOne\'s official mascot was <img class="joypixels" alt="🦄" title=":unicorn:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f984.png"/>!';
+        $image_fix = 'EmojiOne\'s official mascot was <img class="joypixels" alt="&#x1f984;" title=":unicorn:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f984.png"/>!';
 
-        $this->assertEquals(Emojione::toShort($unicode), $shortname);
-        $this->assertEquals(Emojione::shortnameToImage($shortname), $image_fix);
-        $this->assertEquals(Emojione::shortnameToUnicode($shortname), $unicode);
-        $this->assertEquals(Emojione::unicodeToImage($unicode), $image);
-        $this->assertEquals(Emojione::toImage($unicode), $image);
-        $this->assertEquals(Emojione::toImage($shortname), $image_fix);
+        $this->assertEquals(JoyPixels::toShort($unicode), $shortname);
+        $this->assertEquals(JoyPixels::shortnameToImage($shortname), $image_fix);
+        $this->assertEquals(JoyPixels::shortnameToUnicode($shortname), $unicode);
+        $this->assertEquals(JoyPixels::unicodeToImage($unicode), $image);
+        $this->assertEquals(JoyPixels::toImage($unicode), $image);
+        $this->assertEquals(JoyPixels::toImage($shortname), $image_fix);
     }
 
     /**
@@ -160,17 +160,17 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
      */
     public function testShortnameAtEndOfSentenceWithPreceedingColon()
     {
-        $unicode   = 'EmojiOne\'s official mascot: 🦄';
-        $shortname = 'EmojiOne\'s official mascot: :unicorn:';
-        $image     = 'EmojiOne\'s official mascot: <img class="emojione" alt="🦄" title=":unicorn:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f984.png"/>';
-        $image_fix = 'EmojiOne\'s official mascot: <img class="emojione" alt="&#x1f984;" title=":unicorn:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f984.png"/>';
+        $unicode   = 'EmojiOne\'s official mascot was: 🦄';
+        $shortname = 'EmojiOne\'s official mascot was: :unicorn:';
+        $image     = 'EmojiOne\'s official mascot was: <img class="joypixels" alt="🦄" title=":unicorn:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f984.png"/>';
+        $image_fix = 'EmojiOne\'s official mascot was: <img class="joypixels" alt="&#x1f984;" title=":unicorn:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f984.png"/>';
 
-        $this->assertEquals(Emojione::toShort($unicode), $shortname);
-        $this->assertEquals(Emojione::shortnameToImage($shortname), $image_fix);
-        $this->assertEquals(Emojione::shortnameToUnicode($shortname), $unicode);
-        $this->assertEquals(Emojione::unicodeToImage($unicode), $image);
-        $this->assertEquals(Emojione::toImage($unicode), $image);
-        $this->assertEquals(Emojione::toImage($shortname), $image_fix);
+        $this->assertEquals(JoyPixels::toShort($unicode), $shortname);
+        $this->assertEquals(JoyPixels::shortnameToImage($shortname), $image_fix);
+        $this->assertEquals(JoyPixels::shortnameToUnicode($shortname), $unicode);
+        $this->assertEquals(JoyPixels::unicodeToImage($unicode), $image);
+        $this->assertEquals(JoyPixels::toImage($unicode), $image);
+        $this->assertEquals(JoyPixels::toImage($shortname), $image_fix);
     }
 
     /**
@@ -180,15 +180,15 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
      */
     public function testShortnameInsideOfImgTag()
     {
-        $unicode   = 'The <img class="emojione" alt="🦄" title=":unicorn:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f984.png" /> is EmojiOne\'s official mascot.';
-        $shortname = 'The <img class="emojione" alt=":unicorn:" title=":unicorn:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f984.png" /> is EmojiOne\'s official mascot.';
+        $unicode   = 'The <img class="joypixels" alt="🦄" title=":unicorn:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f984.png" /> was EmojiOne\'s official mascot.';
+        $shortname = 'The <img class="joypixels" alt=":unicorn:" title=":unicorn:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f984.png" /> was EmojiOne\'s official mascot.';
 
-        $this->assertEquals(Emojione::toShort($unicode), $unicode);
-        $this->assertEquals(Emojione::shortnameToImage($shortname), $shortname);
-        $this->assertEquals(Emojione::shortnameToUnicode($shortname), $shortname);
-        $this->assertEquals(Emojione::unicodeToImage($unicode), $unicode);
-        $this->assertEquals(Emojione::toImage($unicode), $unicode);
-        $this->assertEquals(Emojione::toImage($shortname), $shortname);
+        $this->assertEquals(JoyPixels::toShort($unicode), $unicode);
+        $this->assertEquals(JoyPixels::shortnameToImage($shortname), $shortname);
+        $this->assertEquals(JoyPixels::shortnameToUnicode($shortname), $shortname);
+        $this->assertEquals(JoyPixels::unicodeToImage($unicode), $unicode);
+        $this->assertEquals(JoyPixels::toImage($unicode), $unicode);
+        $this->assertEquals(JoyPixels::toImage($shortname), $shortname);
     }
 
     /**
@@ -199,25 +199,25 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
     public function testSingleSmiley()
     {
         // enable ASCII conversion
-        $default_ascii = Emojione::$ascii;
-        Emojione::$ascii = true;
+        $default_ascii = JoyPixels::$ascii;
+        JoyPixels::$ascii = true;
 
         $ascii       = ':-)';
         $unicode     = '🙂';
         $unicode_fix = '&#x1f642;';
         $shortname   = ':slight_smile:';
-        $image       = '<img class="emojione" alt="&#x1f642;" title=":slight_smile:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f642.png"/>';
+        $image       = '<img class="joypixels" alt="&#x1f642;" title=":slight_smile:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f642.png"/>';
 
-        $this->assertEquals(Emojione::shortnameToImage($shortname), $image);
-        $this->assertEquals(Emojione::shortnameToImage($ascii), $image);
-        $this->assertEquals(Emojione::toImage($shortname), $image);
-        $this->assertEquals(Emojione::toImage($ascii), $image);
-        $this->assertEquals(Emojione::shortnameToAscii($shortname), ':]');
-        $this->assertEquals(Emojione::unifyUnicode($ascii), $unicode_fix);
-        $this->assertEquals(Emojione::unifyUnicode($shortname), $unicode);
+        $this->assertEquals(JoyPixels::shortnameToImage($shortname), $image);
+        $this->assertEquals(JoyPixels::shortnameToImage($ascii), $image);
+        $this->assertEquals(JoyPixels::toImage($shortname), $image);
+        $this->assertEquals(JoyPixels::toImage($ascii), $image);
+        $this->assertEquals(JoyPixels::shortnameToAscii($shortname), ':]');
+        $this->assertEquals(JoyPixels::unifyUnicode($ascii), $unicode_fix);
+        $this->assertEquals(JoyPixels::unifyUnicode($shortname), $unicode);
 
         // back to default ASCII conversion
-        Emojione::$ascii = $default_ascii;
+        JoyPixels::$ascii = $default_ascii;
     }
 
     /**
@@ -228,17 +228,17 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
     public function testSingleSmileyWithIncorrectCase()
     {
         // enable ASCII conversion
-        $default_ascii = Emojione::$ascii;
-        Emojione::$ascii = true;
+        $default_ascii = JoyPixels::$ascii;
+        JoyPixels::$ascii = true;
 
         $ascii = ':d';
 
-        $this->assertEquals(Emojione::shortnameToImage($ascii), $ascii);
-        $this->assertEquals(Emojione::toImage($ascii), $ascii);
-        $this->assertEquals(Emojione::unifyUnicode($ascii), $ascii);
+        $this->assertEquals(JoyPixels::shortnameToImage($ascii), $ascii);
+        $this->assertEquals(JoyPixels::toImage($ascii), $ascii);
+        $this->assertEquals(JoyPixels::unifyUnicode($ascii), $ascii);
 
         // back to default ASCII conversion
-        Emojione::$ascii = $default_ascii;
+        JoyPixels::$ascii = $default_ascii;
     }
 
     /**
@@ -249,26 +249,26 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
     public function testMultipleSmilies()
     {
         // enable ASCII conversion
-        $default_ascii = Emojione::$ascii;
-        Emojione::$ascii = true;
+        $default_ascii = JoyPixels::$ascii;
+        JoyPixels::$ascii = true;
 
         $ascii       = ';) :p :*';
         $ascii_fix   = ';^) d: :^*';
         $unicode     = '😉 😛 😘';
         $unicode_fix = '&#x1f609; &#x1f61b; &#x1f618;';
         $shortname   = ':wink: :stuck_out_tongue: :kissing_heart:';
-        $image       = '<img class="emojione" alt="&#x1f609;" title=":wink:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f609.png"/> <img class="emojione" alt="&#x1f61b;" title=":stuck_out_tongue:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f61b.png"/> <img class="emojione" alt="&#x1f618;" title=":kissing_heart:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f618.png"/>';
+        $image       = '<img class="joypixels" alt="&#x1f609;" title=":wink:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f609.png"/> <img class="joypixels" alt="&#x1f61b;" title=":stuck_out_tongue:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f61b.png"/> <img class="joypixels" alt="&#x1f618;" title=":kissing_heart:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f618.png"/>';
 
-        $this->assertEquals(Emojione::shortnameToImage($shortname), $image);
-        $this->assertEquals(Emojione::shortnameToImage($ascii), $image);
-        $this->assertEquals(Emojione::toImage($shortname), $image);
-        $this->assertEquals(Emojione::toImage($ascii), $image);
-        $this->assertEquals(Emojione::shortnameToAscii($shortname), $ascii_fix);
-        $this->assertEquals(Emojione::unifyUnicode($ascii), $unicode_fix);
-        $this->assertEquals(Emojione::unifyUnicode($shortname), $unicode);
+        $this->assertEquals(JoyPixels::shortnameToImage($shortname), $image);
+        $this->assertEquals(JoyPixels::shortnameToImage($ascii), $image);
+        $this->assertEquals(JoyPixels::toImage($shortname), $image);
+        $this->assertEquals(JoyPixels::toImage($ascii), $image);
+        $this->assertEquals(JoyPixels::shortnameToAscii($shortname), $ascii_fix);
+        $this->assertEquals(JoyPixels::unifyUnicode($ascii), $unicode_fix);
+        $this->assertEquals(JoyPixels::unifyUnicode($shortname), $unicode);
 
         // back to default ASCII conversion
-        Emojione::$ascii = $default_ascii;
+        JoyPixels::$ascii = $default_ascii;
     }
 
     /**
@@ -279,26 +279,26 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
     public function testSmileyAtSentenceStart()
     {
         // enable ASCII conversion
-        $default_ascii = Emojione::$ascii;
-        Emojione::$ascii = true;
+        $default_ascii = JoyPixels::$ascii;
+        JoyPixels::$ascii = true;
 
         $ascii       = ':\\ is our confused smiley.';
         $ascii_fix   = '=L is our confused smiley.';
         $unicode     = '😕 is our confused smiley.';
         $unicode_fix = '&#x1f615; is our confused smiley.';
         $shortname   = ':confused: is our confused smiley.';
-        $image       = '<img class="emojione" alt="&#x1f615;" title=":confused:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f615.png"/> is our confused smiley.';
+        $image       = '<img class="joypixels" alt="&#x1f615;" title=":confused:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f615.png"/> is our confused smiley.';
 
-        $this->assertEquals(Emojione::shortnameToImage($shortname), $image);
-        $this->assertEquals(Emojione::shortnameToImage($ascii), $image);
-        $this->assertEquals(Emojione::toImage($shortname), $image);
-        $this->assertEquals(Emojione::toImage($ascii), $image);
-        $this->assertEquals(Emojione::shortnameToAscii($shortname), $ascii_fix);
-        $this->assertEquals(Emojione::unifyUnicode($ascii), $unicode_fix);
-        $this->assertEquals(Emojione::unifyUnicode($shortname), $unicode);
+        $this->assertEquals(JoyPixels::shortnameToImage($shortname), $image);
+        $this->assertEquals(JoyPixels::shortnameToImage($ascii), $image);
+        $this->assertEquals(JoyPixels::toImage($shortname), $image);
+        $this->assertEquals(JoyPixels::toImage($ascii), $image);
+        $this->assertEquals(JoyPixels::shortnameToAscii($shortname), $ascii_fix);
+        $this->assertEquals(JoyPixels::unifyUnicode($ascii), $unicode_fix);
+        $this->assertEquals(JoyPixels::unifyUnicode($shortname), $unicode);
 
         // back to default ASCII conversion
-        Emojione::$ascii = $default_ascii;
+        JoyPixels::$ascii = $default_ascii;
     }
 
     /**
@@ -309,26 +309,26 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
     public function testSmileyAtSentenceEnd()
     {
         // enable ASCII conversion
-        $default_ascii = Emojione::$ascii;
-        Emojione::$ascii = true;
+        $default_ascii = JoyPixels::$ascii;
+        JoyPixels::$ascii = true;
 
         $ascii       = 'Our smiley to represent joy is :\')';
         $ascii_fix   = 'Our smiley to represent joy is :\'-)';
         $unicode     = 'Our smiley to represent joy is 😂';
         $unicode_fix = 'Our smiley to represent joy is &#x1f602;';
         $shortname   = 'Our smiley to represent joy is :joy:';
-        $image       = 'Our smiley to represent joy is <img class="emojione" alt="&#x1f602;" title=":joy:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f602.png"/>';
+        $image       = 'Our smiley to represent joy is <img class="joypixels" alt="&#x1f602;" title=":joy:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f602.png"/>';
 
-        $this->assertEquals(Emojione::shortnameToImage($shortname), $image);
-        $this->assertEquals(Emojione::shortnameToImage($ascii), $image);
-        $this->assertEquals(Emojione::toImage($shortname), $image);
-        $this->assertEquals(Emojione::toImage($ascii), $image);
-        $this->assertEquals(Emojione::shortnameToAscii($shortname), $ascii_fix);
-        $this->assertEquals(Emojione::unifyUnicode($ascii), $unicode_fix);
-        $this->assertEquals(Emojione::unifyUnicode($shortname), $unicode);
+        $this->assertEquals(JoyPixels::shortnameToImage($shortname), $image);
+        $this->assertEquals(JoyPixels::shortnameToImage($ascii), $image);
+        $this->assertEquals(JoyPixels::toImage($shortname), $image);
+        $this->assertEquals(JoyPixels::toImage($ascii), $image);
+        $this->assertEquals(JoyPixels::shortnameToAscii($shortname), $ascii_fix);
+        $this->assertEquals(JoyPixels::unifyUnicode($ascii), $unicode_fix);
+        $this->assertEquals(JoyPixels::unifyUnicode($shortname), $unicode);
 
         // back to default ASCII conversion
-        Emojione::$ascii = $default_ascii;
+        JoyPixels::$ascii = $default_ascii;
     }
 
     /**
@@ -339,26 +339,26 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
     public function testSmileyAtSentenceEndWithPunctuation()
     {
         // enable ASCII conversion
-        $default_ascii = Emojione::$ascii;
-        Emojione::$ascii = true;
+        $default_ascii = JoyPixels::$ascii;
+        JoyPixels::$ascii = true;
 
         $ascii       = 'The reverse is the joy smiley is the cry smiley :\'(.';
         $ascii_fix   = 'The reverse is the joy smiley is the cry smiley ;-(.';
         $unicode     = 'The reverse is the joy smiley is the cry smiley 😢.';
         $unicode_fix = 'The reverse is the joy smiley is the cry smiley &#x1f622;.';
         $shortname   = 'The reverse is the joy smiley is the cry smiley :cry:.';
-        $image       = 'The reverse is the joy smiley is the cry smiley <img class="emojione" alt="&#x1f622;" title=":cry:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f622.png"/>.';
+        $image       = 'The reverse is the joy smiley is the cry smiley <img class="joypixels" alt="&#x1f622;" title=":cry:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f622.png"/>.';
 
-        $this->assertEquals(Emojione::shortnameToImage($shortname), $image);
-        $this->assertEquals(Emojione::shortnameToImage($ascii), $image);
-        $this->assertEquals(Emojione::toImage($shortname), $image);
-        $this->assertEquals(Emojione::toImage($ascii), $image);
-        $this->assertEquals(Emojione::shortnameToAscii($shortname), $ascii_fix);
-        $this->assertEquals(Emojione::unifyUnicode($ascii), $unicode_fix);
-        $this->assertEquals(Emojione::unifyUnicode($shortname), $unicode);
+        $this->assertEquals(JoyPixels::shortnameToImage($shortname), $image);
+        $this->assertEquals(JoyPixels::shortnameToImage($ascii), $image);
+        $this->assertEquals(JoyPixels::toImage($shortname), $image);
+        $this->assertEquals(JoyPixels::toImage($ascii), $image);
+        $this->assertEquals(JoyPixels::shortnameToAscii($shortname), $ascii_fix);
+        $this->assertEquals(JoyPixels::unifyUnicode($ascii), $unicode_fix);
+        $this->assertEquals(JoyPixels::unifyUnicode($shortname), $unicode);
 
         // back to default ASCII conversion
-        Emojione::$ascii = $default_ascii;
+        JoyPixels::$ascii = $default_ascii;
     }
 
     /**
@@ -369,26 +369,26 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
     public function testSmileyAtSentenceEndWithPreceedingPunctuation()
     {
         // enable ASCII conversion
-        $default_ascii = Emojione::$ascii;
-        Emojione::$ascii = true;
+        $default_ascii = JoyPixels::$ascii;
+        JoyPixels::$ascii = true;
 
         $ascii       = 'This is the "flushed" smiley: :$.';
         $ascii_fix   = 'This is the "flushed" smiley: =$.';
         $unicode     = 'This is the "flushed" smiley: 😳.';
         $unicode_fix = 'This is the "flushed" smiley: &#x1f633;.';
         $shortname   = 'This is the "flushed" smiley: :flushed:.';
-        $image       = 'This is the "flushed" smiley: <img class="emojione" alt="&#x1f633;" title=":flushed:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f633.png"/>.';
+        $image       = 'This is the "flushed" smiley: <img class="joypixels" alt="&#x1f633;" title=":flushed:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f633.png"/>.';
 
-        $this->assertEquals(Emojione::shortnameToImage($shortname), $image);
-        $this->assertEquals(Emojione::shortnameToImage($ascii), $image);
-        $this->assertEquals(Emojione::toImage($shortname), $image);
-        $this->assertEquals(Emojione::toImage($ascii), $image);
-        $this->assertEquals(Emojione::shortnameToAscii($shortname), $ascii_fix);
-        $this->assertEquals(Emojione::unifyUnicode($ascii), $unicode_fix);
-        $this->assertEquals(Emojione::unifyUnicode($shortname), $unicode);
+        $this->assertEquals(JoyPixels::shortnameToImage($shortname), $image);
+        $this->assertEquals(JoyPixels::shortnameToImage($ascii), $image);
+        $this->assertEquals(JoyPixels::toImage($shortname), $image);
+        $this->assertEquals(JoyPixels::toImage($ascii), $image);
+        $this->assertEquals(JoyPixels::shortnameToAscii($shortname), $ascii_fix);
+        $this->assertEquals(JoyPixels::unifyUnicode($ascii), $unicode_fix);
+        $this->assertEquals(JoyPixels::unifyUnicode($shortname), $unicode);
 
         // back to default ASCII conversion
-        Emojione::$ascii = $default_ascii;
+        JoyPixels::$ascii = $default_ascii;
     }
 
     /**
@@ -399,18 +399,18 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
     public function testSmileyInsideAnImgTag()
     {
         // enable ASCII conversion
-        $default_ascii = Emojione::$ascii;
-        Emojione::$ascii = true;
+        $default_ascii = JoyPixels::$ascii;
+        JoyPixels::$ascii = true;
 
-        $image = 'Smile <img class="emojione" alt=":)" title=":smile:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f642.png" /> because it\'s going to be a good day.';
+        $image = 'Smile <img class="joypixels" alt=":)" title=":smile:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f642.png" /> because it\'s going to be a good day.';
 
-        $this->assertEquals(Emojione::shortnameToImage($image), $image);
-        $this->assertEquals(Emojione::toImage($image), $image);
-        $this->assertEquals(Emojione::shortnameToAscii($image), $image);
-        $this->assertEquals(Emojione::unifyUnicode($image), $image);
+        $this->assertEquals(JoyPixels::shortnameToImage($image), $image);
+        $this->assertEquals(JoyPixels::toImage($image), $image);
+        $this->assertEquals(JoyPixels::shortnameToAscii($image), $image);
+        $this->assertEquals(JoyPixels::unifyUnicode($image), $image);
 
         // back to default ASCII conversion
-        Emojione::$ascii = $default_ascii;
+        JoyPixels::$ascii = $default_ascii;
     }
 
     /**
@@ -421,26 +421,26 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
     public function testTypicalUsernamePasswordFail()
     {
         // enable ASCII conversion
-        $default_ascii = Emojione::$ascii;
-        Emojione::$ascii = true;
+        $default_ascii = JoyPixels::$ascii;
+        JoyPixels::$ascii = true;
 
         $ascii       = 'Please log-in with user:pass as your credentials :P.';
         $ascii_fix   = 'Please log-in with user:pass as your credentials d:.';
         $unicode     = 'Please log-in with user:pass as your credentials 😛.';
         $unicode_fix = 'Please log-in with user:pass as your credentials &#x1f61b;.';
         $shortname   = 'Please log-in with user:pass as your credentials :stuck_out_tongue:.';
-        $image       = 'Please log-in with user:pass as your credentials <img class="emojione" alt="&#x1f61b;" title=":stuck_out_tongue:" src="https://cdn.jsdelivr.net/emojione/assets/' . $this->emojiVersion . '/png/32/1f61b.png"/>.';
+        $image       = 'Please log-in with user:pass as your credentials <img class="joypixels" alt="&#x1f61b;" title=":stuck_out_tongue:" src="https://cdn.jsdelivr.net/joypixels/assets/' . $this->emojiVersion . '/png/32/1f61b.png"/>.';
 
-        $this->assertEquals(Emojione::shortnameToImage($shortname), $image);
-        $this->assertEquals(Emojione::shortnameToImage($ascii), $image);
-        $this->assertEquals(Emojione::toImage($shortname), $image);
-        $this->assertEquals(Emojione::toImage($ascii), $image);
-        $this->assertEquals(Emojione::shortnameToAscii($shortname), $ascii_fix);
-        $this->assertEquals(Emojione::unifyUnicode($ascii), $unicode_fix);
-        $this->assertEquals(Emojione::unifyUnicode($shortname), $unicode);
+        $this->assertEquals(JoyPixels::shortnameToImage($shortname), $image);
+        $this->assertEquals(JoyPixels::shortnameToImage($ascii), $image);
+        $this->assertEquals(JoyPixels::toImage($shortname), $image);
+        $this->assertEquals(JoyPixels::toImage($ascii), $image);
+        $this->assertEquals(JoyPixels::shortnameToAscii($shortname), $ascii_fix);
+        $this->assertEquals(JoyPixels::unifyUnicode($ascii), $unicode_fix);
+        $this->assertEquals(JoyPixels::unifyUnicode($shortname), $unicode);
 
         // back to default ASCII conversion
-        Emojione::$ascii = $default_ascii;
+        JoyPixels::$ascii = $default_ascii;
     }
 
     /**
@@ -451,17 +451,17 @@ class ConversionTest extends \PHPUnit_Framework_TestCase
     public function testSmileyInAnUrl()
     {
         // enable ASCII conversion
-        $default_ascii = Emojione::$ascii;
-        Emojione::$ascii = true;
+        $default_ascii = JoyPixels::$ascii;
+        JoyPixels::$ascii = true;
 
-        $ascii = 'Check out http://www.emojione.com';
+        $ascii = 'Check out http://www.joypixels.com';
 
-        $this->assertEquals(Emojione::shortnameToImage($ascii), $ascii);
-        $this->assertEquals(Emojione::toImage($ascii), $ascii);
-        $this->assertEquals(Emojione::shortnameToAscii($ascii), $ascii);
-        $this->assertEquals(Emojione::unifyUnicode($ascii), $ascii);
+        $this->assertEquals(JoyPixels::shortnameToImage($ascii), $ascii);
+        $this->assertEquals(JoyPixels::toImage($ascii), $ascii);
+        $this->assertEquals(JoyPixels::shortnameToAscii($ascii), $ascii);
+        $this->assertEquals(JoyPixels::unifyUnicode($ascii), $ascii);
 
         // back to default ASCII conversion
-        Emojione::$ascii = $default_ascii;
+        JoyPixels::$ascii = $default_ascii;
     }
 }